Replace floats with integers in sticker struct

+++ src/facelet_cube/moves.rs
@@ -3,14 +3,14 @@ use std::collections::HashMap;
 use crate::generic_cube::{Cube, Move, MoveVariant};
 use crate::geometric_cube::{GeoCube, Sticker, cube3};
 
-pub fn from_geometric_move(mv: Move) -> Vec<(i64, i64)> {
-    let index_map: HashMap<(i64, i64, i64), i64> = create_index_conversion_map();
+pub fn from_geometric_move(mv: Move) -> Vec<(i32, i32)> {
+    let index_map: HashMap<(i32, i32, i32), i32> = create_index_conversion_map();
 
     cube3().apply_move(mv).0
            .iter()
            .map(|s|
-                match (index_map.get(&(s.destination.x as i64, s.destination.y as i64, s.destination.z as i64)),
-                       index_map.get(&(s.position.x as i64, s.position.y as i64, s.position.z as i64))
+                match (index_map.get(&(s.destination.x, s.destination.y, s.destination.z)),
+                       index_map.get(&(s.position.x, s.position.y, s.position.z))
                 ) {
                     (Some(x), Some(y)) => return (x.clone(), y.clone()),
                     (_, _) => panic!()
@@ -20,7 +20,7 @@ pub fn from_geometric_move(mv: Move) -> Vec<(i64, i64)> {
            .collect()
 }
 
-fn create_index_conversion_map() -> HashMap<(i64, i64, i64), i64> {
+fn create_index_conversion_map() -> HashMap<(i32, i32, i32), i32> {
     use Move::*;
     use MoveVariant::*;
 
@@ -39,9 +39,9 @@ fn create_index_conversion_map() -> HashMap<(i64, i64, i64), i64> {
     for rotation in face_rotating_moves {
         for z in [-2, 0, 2] {
             for x in [-2, 0, 2] {
-                let pos = GeoCube(vec![Sticker::new(x as f64, 3 as f64, z as f64)])
+                let pos = GeoCube(vec![Sticker::new(x, 3, z)])
                                 .apply_moves(rotation.clone()).0[0].position;
-                map.insert((pos.x as i64, pos.y as i64, pos.z as i64), idx);
+                map.insert((pos.x, pos.y, pos.z), idx);
                 idx += 1;
             }
         }
